Added new custom.js file and user permissions
I created a new "karaoke.admin.js" file and added key permissions in
registering users.

+++ delete_video.php
@@ -2,39 +2,45 @@
 
 	require_once 'includes/init.php';
 	
-	if (!Input::exists('get')) {
-		Redirect::to('../index.php');
-	}
+	$user = new User();
 	
-	foreach($_GET as $key => $value){
-		if ($value === ""){
+	if (!$user->isLoggedIn()){
+		Redirect::to('index.php');
+	} else {
+		if (!Input::exists('get')) {
 			Redirect::to('../index.php');
 		}
-	}
 		
-	if (Input::exists('get')){
-		$video = new Video();
-		$video_id = $video->safe_string(Input::get('video_id'));
-		$video->find($video_id);
-		$song_title = $video->safe_string(Input::get('song_title'));
-	}
-	
-	
-	
-	if (!$video->exists()){
-		Redirect::to('../index.php');
-	} else {
-		//echo "Hello World!";
-		try{
-			$video->delete([
-				'video_id', '=' , Input::get('video_id')
-			]);
-			
-			Session::flash('delete', 'Video for song "{$song_title}" has been deleted from the database.');
-			//Redirect::to('../index.php');
-		} catch (Exception $e){
-			die($e->getMessage());
+		foreach($_GET as $key => $value){
+			if ($value === ""){
+				Redirect::to('../index.php');
+			}
 		}
-		//print_r($video->data());
+			
+		if (Input::exists('get')){
+			$video = new Video();
+			$video_id = $video->safe_string(Input::get('video_id'));
+			$video->find($video_id);
+			$song_title = $video->safe_string(Input::get('song_title'));
+		}
+		
+		
+		
+		if (!$video->exists()){
+			Redirect::to('../index.php');
+		} else {
+			//echo "Hello World!";
+			try{
+				$video->delete([
+					'video_id', '=' , Input::get('video_id')
+				]);
+				
+				Session::flash('delete', 'Video for song "{$song_title}" has been deleted from the database.');
+				//Redirect::to('../index.php');
+			} catch (Exception $e){
+				die($e->getMessage());
+			}
+			//print_r($video->data());
+		}
 	}
 ?>
